+++ plugins/google-analytics-dashboard/src/components/TimeRanges/index.tsx
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { default } from './TimeRange';
+export { default } from './TimeRanges';
